requirements.txt
aiogram>=3.24.0
mutagen>=1.47.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
Pillow>=10.0.0
fastapi>=0.109.0
uvicorn>=0.27.0